+++ README.md
@@ -44,4 +44,4 @@ flutterfire configure
 
 ## Demo
 
-![chatting](https://github.com/reddtsai/issue/blob/main/images/reddslife-chatting.gif) ![notification](https://github.com/reddtsai/issue/blob/main/images/reddslife-notification.png)
+![chatting](https://github.com/reddtsai/issue/blob/main/images/reddslife-chatting.gif | width=250) ![notification](https://github.com/reddtsai/issue/blob/main/images/reddslife-notification.png | width=250)
