+++ lib/controllers/chats_controller.dart
@@ -1,6 +1,6 @@
 import 'package:get/get.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import '../data/firebase_service.dart';
+import '../services/firebase_service.dart';
 
 class ChatsController extends GetxController {
   final FirebaseService firebaseService;
